从本地服务端以及环信删除联系人

+++ src/cn/ucai/superwechat/activity/MainActivity.java
@@ -587,19 +587,37 @@ public class MainActivity extends BaseActivity implements EMEventListener {
 			// 刷新ui
 			if (currentTabIndex == 1)
 				contactListFragment.refresh();
-
 		}
-
 		@Override
 		public void onContactDeleted(final List<String> usernameList) {
 			// 被删除
 			Log.e(TAG, "onContactDeleted=" + usernameList);
 
-			Map<String, User> localUsers = ((DemoHXSDKHelper)HXSDKHelper.getInstance()).getContactList();
-			for (String username : usernameList) {
-				localUsers.remove(username);
-				userDao.deleteContact(username);
-				inviteMessgeDao.deleteMessage(username);
+			for (final String username : usernameList) {
+				Log.e(TAG, "username=" + username);
+				final OkHttpUtils2<Result> utils2 = new OkHttpUtils2<>();
+				utils2.setRequestUrl(I.REQUEST_DELETE_CONTACT)
+						.addParam(I.Contact.USER_NAME,SuperWeChatApplication.getInstance().getUserName())
+						.addParam(I.Contact.CU_NAME,username)
+						.targetClass(Result.class)
+						.execute(new OkHttpUtils2.OnCompleteListener<Result>() {
+							@Override
+							public void onSuccess(Result result) {
+								if (result.isRetMsg()) {
+									UserAvatar userAvatar = SuperWeChatApplication.getInstance().getUserAvatarMap().get(username);
+									SuperWeChatApplication.getInstance().getUserList().remove(userAvatar);
+									SuperWeChatApplication.getInstance().getUserAvatarMap().remove(username);
+									userDao.deleteContact(username);
+									((DemoHXSDKHelper)HXSDKHelper.getInstance()).getContactList().remove(username);
+									inviteMessgeDao.deleteMessage(username);
+									sendBroadcast(new Intent("update_contact_list"));
+								}
+							}
+							@Override
+							public void onError(String error) {
+
+							}
+						});
 			}
 			runOnUiThread(new Runnable() {
 				public void run() {
@@ -607,7 +625,7 @@ public class MainActivity extends BaseActivity implements EMEventListener {
 					String st10 = getResources().getString(R.string.have_you_removed);
 					if (ChatActivity.activityInstance != null
 							&& usernameList.contains(ChatActivity.activityInstance.getToChatUsername())) {
-						Toast.makeText(MainActivity.this, ChatActivity.activityInstance.getToChatUsername() + st10, 1)
+						Toast.makeText(MainActivity.this, ChatActivity.activityInstance.getToChatUsername() + st10, Toast.LENGTH_LONG)
 								.show();
 						ChatActivity.activityInstance.finish();
 					}
